NPCs now give hints to where important items are!
The security guard will tell the player where the vault keycard is. (On his person / In the security room / In the bank manager's office)
The bank manager will tell the player where the vault code memo is. (If there is one) (On his person / In the safe at his office)